moving publipostage code from Controller to Service
